Update border overrides
Properly reporioritizes the layer based on settings. Separates the
border fill to a ne layer that is always on the bottom and always drawn
if enabled.  Wraps the getShape method to make gridless tokens round.

Notably, removes all special directives for hex tokens. Foundry handles
the border shape for almost every token shape, and handles orientation.